+++ uamp-client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
 - Add progress bar dragging
+- Add song context menu
 - Fix volume overwrite on startup
 - Fix album hover effect
+- Fix queue and play-next event handling